categorizes every util function
