+++ src/js/util/array/index.ts
@@ -1,0 +1,1 @@
+export * from './toggle-item-in-array';
