+++ src/js/util/array/toggle-item-in-array.ts
@@ -1,0 +1,2 @@
+export const toggleItemInArray = (item, arr) =>
+  arr.includes(item) ? arr.filter(k => k !== item) : [...arr, item];
