+++ src/js/util/misc/is-falsy.ts
@@ -1,0 +1,2 @@
+export const isFalsy = value =>
+  [null, 'null', '', undefined, 'undefined'].some(v => value === v);
