+++ src/js/util/misc/object-to-query-params.ts
@@ -1,0 +1,5 @@
+export const objectToQueryParams = object =>
+  Object.entries(object)
+    .filter(([, value]) => [undefined, NaN, null, ''].every(n => n !== value))
+    .map(([key, value]) => `${key}=${encodeURIComponent(value)}`)
+    .join('&');
