+++ src/js/util/url/to-search-url.ts
@@ -1,18 +1,15 @@
-import { objectToQueryParams } from '..';
-
-/* globals API_URL */
 import {
   DEFAULT_SEARCH_TYPE,
   DEFAULT_SEARCH_SOURCE,
 } from '../../constants';
 
+import { objectToQueryParams } from '../misc';
 interface IToSearchURLArguments {
   query: string,
   type: number,
   source: string,
   offset: string
 }
-
 
 export const toSearchURL = ({
   query: q,
